requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
Pillow>=10.0.0
edge-tts>=6.1.9
requests-oauthlib>=1.3.1
